requirements.txt
asgiref==3.8.1
Django==3.2.3
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-js-asset==2.2.0
django-tinymce==3.3.0
djangorestframework==3.12.4
Pillow==8.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
